Move ProcedureNotUsedInspectionTests for QuickFixes into relevant files under RD.QuickFixes
